Updated flatpacks

+++ warehouse/moondream/custom.sh
@@ -3,5 +3,9 @@
 # "${VENV_PIP}"
 # "${VENV_PYTHON}"
 
+init_script="$SCRIPT_DIR/init.sh"
+[ -f "$init_script" ] || { echo "init.sh not found, exiting."; exit 1; }
+source "$init_script" || { echo "Failed to load init.sh."; exit 1; }
+
 cp ../tiger.png tiger.png
 "${VENV_PYTHON}" sample.py --image "tiger.png" --prompt "Should I pet this dog?"
